Some minor refactoring

+++ src/brute.h
@@ -4,7 +4,9 @@
 #include <fstream>
 #include <windows.h>
 
-const DWORD THREAD_COUNT_DEFAULT = 4;
+#include "utils.h"
+
+const ulong THREAD_COUNT_DEFAULT = 4;
 
 void brute(int argc, char** argv, ifstream* fp);
